feat: create team_stats module and corresponding testing
