+++ spec/stat_tracker_spec.rb
@@ -106,57 +106,4 @@ RSpec.describe StatTracker do
     expect(@stat_tracker.lowest_scoring_home_team).to eq('New York City FC')
   end
 
-  it '#team_info' do
-    expect(@stat_tracker.team_info('18')).to eq({
-      "team_id" => "18",
-      "franchise_id" => "34",
-      "team_name" => "Minnesota United FC",
-      "abbreviation" => "MIN",
-      "link" => "/api/v1/teams/18"
-    })
-  end
-
-  it '#total_wins_per_season' do
-    expect(@stat_tracker.total_wins_per_season('6')).to eq({"20122013"=>2})
-  end
-
-  it '#total_games_played_per_season' do
-    expect(@stat_tracker.total_games_played_per_season('6')).to eq({"20122013"=>4})
-  end
-
-  it '#best_season' do
-    expect(@stat_tracker.best_season('6')).to eq("20122013")
-  end
-
-  it '#worst_season' do
-    expect(@stat_tracker.worst_season('6')).to eq("20122013")
-  end
-
-  it '#average_win_percentage' do
-    expect(@stat_tracker.average_win_percentage('6')).to eq(0.5)
-  end
-
-  it '#most_goals_scored' do
-    expect(@stat_tracker.most_goals_scored('8')).to eq(3)
-  end
-
-  it "#fewest_goals_scored" do
-    expect(@stat_tracker.fewest_goals_scored('16')).to eq(0)
-  end
-
-  it '#total_times_won_against' do
-    expect(@stat_tracker.total_times_won_against('6')).to eq({3=>2})
-  end
-
-  it '#total_times_played_against' do
-    expect(@stat_tracker.total_times_played_against('6')).to eq({3=>4})
-  end
-
-  it '#favorite_opponent' do
-    expect(@stat_tracker.favorite_opponent('6')).to eq("Houston Dynamo")
-  end
-
-  it '#rival' do
-    expect(@stat_tracker.rival('6')).to eq('Houston Dynamo')
-  end
 end
